requirements.txt
python-telegram-bot
django
python-dotenv
requests
pandas
xlrd
datetime
xlutils